requirements.txt
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
pytest>=7.4.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

fastapi
uvicorn
httpx

# Optional: GPU acceleration (requires CUDA)
# cupy-cuda12x
